requirements.txt
pandas
opencv-python
pdf2image
paddlepaddle
paddleocr
Pillow
